requirements.txt
blinker==1.3
coverage==4.0.2
docutils==0.12
Flask==0.10.1
Flask-DebugToolbar==0.10.0
Flask-SQLAlchemy==2.0
funcsigs==0.4
geolocation-python==0.2.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
mock==1.3.0
pbr==1.8.1
psycopg2==2.6.1
requests==2.8.1
six==1.10.0
SQLAlchemy==1.0.3
Werkzeug==0.10.4
wheel==0.24.0
